fix(sla): cast Carbon diffIn*() results to int at correct-direction call sites
1A of the Carbon 2->3 sweep. Carbon 3 changed diffIn*() to return a
signed float (was an unsigned int). These call sites already pass
start->diff(end) in the right order, so the value stays correct — the
only visible change was spurious decimals ("45.3 วิ") and float sums.
Wrapping each in (int) restores the pre-upgrade output on well-formed
data; no logic/direction change.

Sites:
- SlaPerformanceController: response/acceptance/resolution minute sums
  (lines ~72, 76, 81, 131, 201) + chart month span guard (~184)
- sla/index.blade.php: date-range day count (~190/195), breached
  overdue minutes (~481), at-risk remaining minutes (~542)
- _timeline.blade.php: per-state duration seconds (~48)
- _sla_info.blade.php: response / live-hold / gross-repair seconds
  (~9, 14, 30)

Verified: php -l clean, `artisan view:cache` compiles all Blade,
test suite unchanged (22 pass / 6 pre-existing fail), and SLA dashboard
+ request-show SLA panel render with clean integer durations.

Not touched here (need direction/abs fixes — item 1B, pending sign-off):
MaintenanceRatingController:304, Api/MaintenanceRatingApiController:144,
MaintenanceJobController:131.

+++ resources/views/maintenance/sla/index.blade.php
@@ -187,10 +187,10 @@
                         // คำนวณจำนวนวัน
                         $diffDays = 0;
                         if ($fromDate) {
-                            $diffDays = $fromDate->startOfDay()->diffInDays($toDate->copy()->startOfDay()) + 1;
+                            $diffDays = (int) $fromDate->startOfDay()->diffInDays($toDate->copy()->startOfDay()) + 1;
                         } else {
                             $diffDays =
-                                now()
+                                (int) now()
                                     ->startOfYear()
                                     ->diffInDays(now()->startOfDay()) + 1;
                         }
@@ -478,7 +478,7 @@
                                 @forelse (collect($breachedTickets)->take(20) as $t)
                                     @php
                                         $now = \Carbon\Carbon::now();
-                                        $diffInMins = $t->sla_due_date->diffInMinutes($now);
+                                        $diffInMins = (int) $t->sla_due_date->diffInMinutes($now);
                                         $days = floor($diffInMins / (60 * 24));
                                         $hrs = floor(($diffInMins % (60 * 24)) / 60);
                                         $mins = $diffInMins % 60;
@@ -539,7 +539,7 @@
                                 @forelse (collect($atRiskTickets)->take(20) as $t)
                                     @php
                                         $now = \Carbon\Carbon::now();
-                                        $diffInMins = $now->diffInMinutes($t->sla_due_date);
+                                        $diffInMins = (int) $now->diffInMinutes($t->sla_due_date);
                                         // 4 hours warning baseline (total = 240 mins)
                                         $pct = min(100, max(0, 100 - ($diffInMins / (4 * 60)) * 100));
                                         $hrs = floor($diffInMins / 60);
